+++ src/test/java/it/pagopa/pn/paperchannel/config/BaseTest.java
@@ -1,12 +1,18 @@
 package it.pagopa.pn.paperchannel.config;
 
+import io.awspring.cloud.autoconfigure.messaging.SqsAutoConfiguration;
 import it.pagopa.pn.paperchannel.LocalStackTestConfig;
+import it.pagopa.pn.paperchannel.middleware.queue.producer.DeliveryPushMomProducer;
+import it.pagopa.pn.paperchannel.middleware.queue.producer.InternalQueueMomProducer;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.cloud.function.context.config.ContextFunctionCatalogAutoConfiguration;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
@@ -21,11 +27,18 @@ public abstract class BaseTest {
 
     @Slf4j
     @SpringBootTest
+    @EnableAutoConfiguration(exclude= {SqsAutoConfiguration.class, ContextFunctionCatalogAutoConfiguration.class})
     @ActiveProfiles("test")
-    @Import(LocalStackTestConfig.class)
     public static class WithMockServer {
         @Autowired
         private MockServerBean mockServer;
+
+        @MockBean
+        private DeliveryPushMomProducer deliveryMomProducer;
+
+        @MockBean
+        private InternalQueueMomProducer internalQueueMomProducer;
+
         @BeforeEach
         public void init(){
             log.info(this.getClass().getSimpleName());
