+++ src/test/java/it/pagopa/pn/paperchannel/middleware/msclient/ExternalChannelClientTest.java
@@ -13,7 +13,6 @@ import org.springframework.web.reactive.function.client.WebClientResponseExcepti
 
 import java.time.Instant;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 class ExternalChannelClientTest  extends BaseTest.WithMockServer {
