+++ src/test/java/it/pagopa/pn/paperchannel/service/PaperListServiceTest.java
@@ -1,15 +1,16 @@
 package it.pagopa.pn.paperchannel.service;
 
-import it.pagopa.pn.paperchannel.config.BaseTest;
 import it.pagopa.pn.paperchannel.generated.openapi.server.v1.dto.CapResponseDto;
 import it.pagopa.pn.paperchannel.middleware.db.dao.CapDAO;
 import it.pagopa.pn.paperchannel.middleware.db.entities.PnCap;
+import it.pagopa.pn.paperchannel.service.impl.PaperListServiceImpl;
 import it.pagopa.pn.paperchannel.utils.Const;
-import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.mockito.junit.jupiter.MockitoExtension;
 import reactor.core.publisher.Mono;
 
 import java.util.ArrayList;
@@ -18,13 +19,13 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-@Slf4j
-class PaperListServiceTest extends BaseTest {
+@ExtendWith(MockitoExtension.class)
+class PaperListServiceTest {
 
-    @MockBean
+    @Mock
     private CapDAO capDAO;
-    @Autowired
-    private PaperListService paperListService;
+    @InjectMocks
+    private PaperListServiceImpl paperListService;
 
 
     @Test
