+++ src/test/java/it/pagopa/pn/paperchannel/service/PrepareAsyncServiceTest.java
@@ -1,6 +1,6 @@
 package it.pagopa.pn.paperchannel.service;
 
-import it.pagopa.pn.paperchannel.config.BaseTest;
+import it.pagopa.pn.paperchannel.config.PnPaperChannelConfig;
 import it.pagopa.pn.paperchannel.exception.ExceptionTypeEnum;
 import it.pagopa.pn.paperchannel.exception.PnAddressFlowException;
 import it.pagopa.pn.paperchannel.exception.PnGenericException;
@@ -13,7 +13,6 @@ import it.pagopa.pn.paperchannel.middleware.db.dao.RequestDeliveryDAO;
 import it.pagopa.pn.paperchannel.middleware.db.entities.PnAddress;
 import it.pagopa.pn.paperchannel.middleware.db.entities.PnAttachmentInfo;
 import it.pagopa.pn.paperchannel.middleware.db.entities.PnDeliveryRequest;
-import it.pagopa.pn.paperchannel.middleware.db.entities.PnRequestError;
 import it.pagopa.pn.paperchannel.middleware.msclient.AddressManagerClient;
 import it.pagopa.pn.paperchannel.middleware.msclient.SafeStorageClient;
 import it.pagopa.pn.paperchannel.model.Address;
@@ -22,9 +21,11 @@ import it.pagopa.pn.paperchannel.service.impl.PrepareAsyncServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.mockito.junit.jupiter.MockitoExtension;
 import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
@@ -39,24 +40,28 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class PrepareAsyncServiceTest extends BaseTest {
+@ExtendWith(MockitoExtension.class)
+class PrepareAsyncServiceTest {
 
-    @Autowired
+    @InjectMocks
     private PrepareAsyncServiceImpl prepareAsyncService;
-    @MockBean
+    @Mock
     private AddressDAO addressDAO;
-    @MockBean
+    @Mock
     private RequestDeliveryDAO requestDeliveryDAO;
-    @MockBean
+    @Mock
     private SqsSender sqsSender;
-    @MockBean
+    @Mock
     private SafeStorageClient safeStorageClient;
-    @MockBean
+    @Mock
     private AddressManagerClient addressManagerClient;
-    @MockBean
+    @Mock
     private PaperAddressService paperAddressService;
-    @MockBean
+    @Mock
     private PaperRequestErrorDAO paperRequestErrorDAO;
+
+    @Mock
+    private PnPaperChannelConfig pnPaperChannelConfig;
 
     private final PrepareAsyncRequest request = new PrepareAsyncRequest();
     private final Address address = new Address();
@@ -114,9 +119,6 @@ class PrepareAsyncServiceTest extends BaseTest {
 
         Mockito.when(this.paperAddressService.getCorrectAddress(Mockito.any(), Mockito.any(), Mockito.any()))
                 .thenThrow(new PnGenericException(UNTRACEABLE_ADDRESS, UNTRACEABLE_ADDRESS.getMessage()));
-
-        Mockito.when(this.paperRequestErrorDAO.created(Mockito.any(), Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.just(new PnRequestError()));
 
         Mockito.when(this.requestDeliveryDAO.getByCorrelationId(Mockito.any()))
             .thenReturn(Mono.just(getDeliveryRequest()));
